Added RefreshIndicator to YoutubePage and the videos will now open in the Youtube App on tap
