+++ lib/widgets/base_page.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
 import 'package:influx/config.dart';
 import 'package:influx/widgets/rss_feed_page.dart';
-import 'package:influx/widgets/youtube_page/youtube_page.dart';
 import 'package:influx/widgets/youtube_page/youtube_page_clone.dart';
 
 class BasePage extends StatefulWidget {
